split hooks for reusability
