+++ client/a8c-for-agencies/sections/referrals/hooks/use-get-consolidated-payout-data.ts
@@ -1,7 +1,6 @@
 import { useMemo } from 'react';
 import { APIProductFamilyProduct } from 'calypso/state/partner-portal/types';
 import { getEstimatedCommission } from '../lib/get-estimated-commission';
-import { getNextPayoutDate, getNextPayoutDateActivityWindow } from '../lib/get-next-payout-date';
 import { Referral } from '../types';
 
 export default function useGetConsolidatedPayoutData(
@@ -23,34 +22,8 @@ export default function useGetConsolidatedPayoutData(
 		[ referrals ]
 	);
 
-	const nextPayoutActivityWindow = useMemo( () => {
-		const { start, finish } = getNextPayoutDateActivityWindow( new Date() );
-		const startDate = start.toLocaleString( 'default', {
-			month: 'short',
-			day: 'numeric',
-			year: 'numeric',
-		} );
-		const finishDate = finish.toLocaleString( 'default', {
-			month: 'short',
-			day: 'numeric',
-			year: 'numeric',
-		} );
-		return `${ startDate } - ${ finishDate }`;
-	}, [] );
-
-	const nextPayoutDate = useMemo(
-		() =>
-			getNextPayoutDate( new Date() ).toLocaleString( 'default', {
-				month: 'short',
-				day: 'numeric',
-			} ),
-		[]
-	);
-
 	return {
 		expectedCommission,
 		pendingOrders,
-		nextPayoutActivityWindow,
-		nextPayoutDate,
 	};
 }
